Potrebno poslati neku update poruku nakon sto se postavi prethodnik na nasem sledbeniku i sledbenik na nasem prethodniku.

+++ src/sillygit/util/FileUtils.java
@@ -245,7 +245,7 @@ public class FileUtils {
                 }
                 dirPath = rootDirectory + "\\" + dirPath;
                 File dir = new File(dirPath);
-                if (dir.listFiles().length == 0) {
+                if (dir != null && dir.listFiles() != null && dir.listFiles().length == 0) {
                     dir.delete();
                 }
             }
